sm5803: Force BFET on when charging from AUX port
In order to have both our charger ICs behave the same way when running
OCPC, this commit changes our usage of the charger IC to force the
BFET to be on during the majority of the charge curve.  This also
improves efficiency over using the previous linear mode.

Additionally, as part of the charger init, the initial VSYS voltage is
set to the battery voltage.

BUG=b:168253008,b:168252106
BRANCH=None
TEST=Build and flash waddledee.  Verify DUT can charge from both
ports.  Verify DUT can source out of both ports.  Verify that DUT can
sink and charge simultaneously from both ports.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I56a59b4aaff11b6f2f190695c8497b19a9dbf9e8