ocpc: Add visualization graph
When tuning an OCPC system, the OCPC debug output is very useful in
order to examine the behaviour of the system.  However, instead of
simply looking at a bunch of changing numbers on the terminal, it's
easier to view a waveform of the current entering the battery.  In
lieu of using an oscilloscope with a current probe, this commit adds a
visualization graph which shows the current entering the battery
relative to the target current.  Each segment on the graph represents
5% of the target current.

Usage is as follows:

    > ocpcdebug enable/viz/all/disable

`enable` will show text debug only.
`viz` will show the visualization graph (output shown below).
`all` will show both the graph and text debug data.
`disable` will show no debug output.

Here is an example of the output

[2665.199505 Act Chg: 1]
[2665.702271 Battery 47% (Display 47.1 %) / ??h:?? to empty]
[2665.703810 charge_request(13050mV, 1800mA)]
[2665.725854 OCPC: Target VSYS: 11979mV]
[----------#----------] (actual)251mA (desired)1880mA
[2666.317342 OCPC: Target VSYS: 12179mV]
[----------|---------o] (actual)879mA (desired)1880mA
[2666.908941 OCPC: Target VSYS: 12274mV]
[----------|-----o----] (actual)1216mA (desired)1880mA
[----------|o---------] (actual)1631mA (desired)1880mA
[2668.090373 OCPC: Target VSYS: 12259mV]
[----------#----------] (actual)1797mA (desired)1880mA
[----------#----------] (actual)1813mA (desired)1880mA
[----------#----------] (actual)1769mA (desired)1880mA
[----------#----------] (actual)1779mA (desired)1880mA

BUG=None
BRANCH=None
TEST=Build and flash waddledee, enable debug command, verify that
visualization graph is shown.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Ie271524ce2ce90781b966a77f2a7ef27dc969075