TBT:Add support for exit mode SOP'/SOP'' on chipset transition
When chipset is transitioning to a new state or on sysjump, the SOC
re-negotiates to enter an alternate mode on booting up.

This commit adds support for exiting Thunderbolt mode for SOP' and SOP''
on chipset transition and also moves all the alternate mode exit rotines
to their respective files.

It also delays deleting the SVID data until after the EXIT_MODE message has
ACKed.

BUG=b:151169925, b:159717794
BRANCH=none
TEST=Able to exit and re-enter into DisplayPort mode, Thunderbolt mode
     with passive cable and thunderbolt mode with active cable cable on
    reboot.

Change-Id: If1e48e9f31cd678e23fe89bd3494551b5d1a78f1
Signed-off-by: Ayushee <ayushee.shah@intel.com>