cortex-m: support MPU addrs less aligned than size
Data RAM on the NPCX797WB is less aligned that the size of that memory
region (0x200B0000 vs 0x20000), which means MPU setup fails because the
current approach for MPU configuration assumes that addr is at least as
aligned as size is.

To make it work on this particular chip, add the ability to configure
two smaller regions each of half size as long as the size is already a
power of two. This allows MPU configuration to succeed as long as the
address is no more than one bit less aligned than the size, which is
sufficient for NPCX797FC and NPCX797WC.

BUG=b:168732993
TEST=Not yet
BRANCH=zork

Change-Id: Ifd31f10ed24b8b777839031c01a47c7454605232
Signed-off-by: Peter Marheine <pmarheine@chromium.org>