common: body_detection: modify calculation of noise level
We used rms_noise^2 as noise level for body_detection. But, there are
still some different noise level based on different devices. Add the
CONFIG_BODY_DETECTION_VAR_NOISE_FACTOR to adjust the noise_level.
Default value is 120%, since the 100% of the rms_noise^2 is not enough,
especially in 200Hz.

BRANCH=None
BUG=b:123434029
TEST=make buildall; make BOARD=ampton, re-flash the DUT, see if the body
     detection works

Signed-off-by: Ching-Kang Yen <chingkang@chromium.org>
Change-Id: I2a7416b0bd8d0aac51d16c3aed9d962837f52c0b