USB-PD: Continue mode discovery after failure
pd_get_next_mode was searching the svid_mode_data
looking for the next entry to discover, if there
are any. Failed and needs discovery were handled
the same way, return that entry. So if you had a
discovery entry that fails that gets returned and
causes the entry to stop, over and over again

This issue was found using an ikling hub with
attached AC connected to Morphius the SVID
discovery finds two SVIDs, DP and Nintendo.  The
Nintendo fails and this caused the alt-mode to
not enter for the valid DP mode.

The change is to continue discovery even if there
is a failure discovered. If any of the modes
succeeded then it will be considered success if
there are no new modes to discover.

This is the order of returned values:
  no entries -> successful.
  an entry that need to be discovered -> return discovery entry.
  everything discovered properly in the list -> successful.
  discovered some good and some failed -> successful.
  all discovered fail -> return first fail.

BUG=b:168804124
BRANCH=none
TEST=verify morphius monitor when attached ikling powered hub

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: Id2064079614c276f3d2054c5bcb194cdb997514c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2416990
Tested-by: Denis Brockus <dbrockus@chromium.org>
Reviewed-by: Denis Brockus <dbrockus@chromium.org>
Commit-Queue: Denis Brockus <dbrockus@chromium.org>
Auto-Submit: Denis Brockus <dbrockus@chromium.org>