USB-PD: AltMode fails if any discovery fails
pd_get_next_mode was searching the svid_mode_data
looking for the next entry to discover, if there
are any. Failed and needs discovery were handled
the same way, return that entry. So if you had a
discovery entry that fails that gets returned and
causes the entry to stop, over and over again

Using an Ikling hub with attached AC connected to
Morphius the SVID discovery finds two SVIDs, DP
and Nintendo.  The Nintendo fails and this causes
the alt-mode to not enter for the valid DP mode.

BUG=b:168804124
BRANCH=none
TEST=verify morphius monitor when attached ikling powered hub

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: Id2064079614c276f3d2054c5bcb194cdb997514c