it83xx: Only consider ITE C ports for sleep mask
The TCPMvX frameworks do not clear the sleep mask for USB PD if the
device is using on "off chip" TCPC like the one embedded in the ITE
EC.  Therefore, the ITE driver would take care of this.

This commit fixes a bug in the ITE TCPC driver which assumed that
boards which are using the ITE TCPC would be using it for all ports.
This caused the driver to never allow deep sleep even if no devices
were connected.

BUG=b:168783892
BRANCH=None
TEST=Build and flash drawcia.  Unplug all devices and verify that the
sleep mask is 0 when the AP is off.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I312f0f2a24d2783291f3c52612e62bc61e4ccca2