servo_v4p1: convert the TUSB1064 driver to a full usb_mux implementation
The TUSB1064 driver code is modified to implement a real usb_mux driver.
This part is a UFP side (RX Equalizer) USB-C switch contrary to most
other usb_mux part we have (used in the DFP).

It is used in the servo_v4p1, re-configure the board code accordingly

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=b:168621142
BRANCH=servo
TEST=use the sink mode enabled by 'cc pdsnk', connect to a USB-C Chromebook
try both video output on USB-C->mini-DP and USB 3.0.

Change-Id: Ie15d246ba7b6601db0d387e8ffda54f0ec36ba99