ServoV4p1: Add hook function to set UFP SSUSB mux
This commit is part of a series to clean up ServoV4p1 code.

TUSB1064 is intended as a UFP endpoint mux. However TCPMv1
mux architecture seems to rely on the assumption that we
are always a DFP. Mux enable calls are only engaged in DFP
operation.

(TCPMv2 does have an independent state machine for DFP.)

Since TCPMv1 is no longer actively being refactored, we
instead attempt to use a HOOK function to configure the
UFP mux as appropriate.

NOTE:
This patch does not cover corner cases such as DTS
operation. This is intended as a "QuickFix" to unblock
testing and validation.

BUG=b/168621142
BRANCH=none
TEST=Compile multi patch series. Look for SSUSB hub
enumeration in CB when in 'cc pdsnk' or 'cc drp' modes.

Change-Id: I3a5d8aa0bc666f3c498b017751fd047e422b2649