npcx7: set code RAM size to actual code RAM size
Munging the program flash size to account for the ROM header by default
also affects the assumed size of code RAM, which causes problems with
the MPU because it makes the code RAM region a weird size. Since we're
not using all of the code RAM space but it remains a full 256k in size,
explicitly set the code RAM size for the chips where we pretend program
memory is smaller than it actually is.

BUG=b:168732993
TEST=verified dalboz configures code RAM with size 256k
BRANCH=zork

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I0ab0954bc8a433d0d7c0c85945b012d45cc225ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2417788
Reviewed-by: Andrew McRae <amcrae@chromium.org>
(cherry picked from commit 3a2d3b3d140c07863f4dcb919f5d7b03505a50ed)