u2f: Switch FIPS call back to DCRYPTO for 5.7/6.7
There is a fips_rand_bytes() call in u2f. Since 5.7/6.7 will be
released without FIPS support (and the plan is to move u2f to FIPS
later), change that call back to DCRYPTO_ladder_random.

This does not affect the canonical (non-versioned) u2f. This only
affects the generation of versioned KHs for WebAuthn purposes.

BUG=none
TEST=make -j BOARD=cr50

Signed-off-by: Yicheng Li <yichengli@chromium.org>
Change-Id: I78142efd1b3a8339cce44adc4d3f8e26151b30ce