host_event_commands: Return access denied for prohibited access
Currently, EC_CMD_HOST_EVENT returns the same error code for
non-existing masks and prohibited access to the existing masks.

This patch makes the command return ACCESS_DENIED for the latter case
so that the host can distinguish the two cases.

ToT BIOS won't be affected because the EC's return codes are collapsed
to -1 or -(request size).

$ ectool hostevent set 1 0
Set isn't permitted for mask 1.

BUG=b:168939843
BRANCH=None
TEST=Verified on Atlas. See above.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: Ia346dbffd459985d5eea8a955e22822d402d5388