tcpm: Mitigate EC RX circular buffer overflow tight-loop
There are two FIFO buffers for USB-PD packet reception. One is located
in the TCPC, another was added in EC layer in crrev/c/1185727 . There
are scenarios that lead to extremely rapid packet reception.

In these scenarios, the EC RX circular buffer overflows without clearing
the TCPC ALERT# signal. This in turn results in a tight-loop in the TCPC
interrupt handler, leading to port "suspend".

This CL does not improve the fundamental issue of message reception in
excess of consumption rate. It only addresses the tight-loop.

Additional work is needed to implement full handling of rapid messages
(permitted in PD3) and logical VESA DP AltMode PD-to-HPD enqueueing.

BRANCH=none
BUG=b:155517411
TEST=1. Connect Acer CB282K via DP to HP G5 dock
     2. Connect HP G5 dock to AKEMI on TCPMv1
     3. Allow MST VDM:Attention communication to occur
     4. Observe EC console prints via SuzyQ

Change-Id: If2459235c9422cf6caa4c38c5101d06d37900dd8
Signed-off-by: Nathan K <nkolluru@google.com>