TCPMv2: Add TYPEC_STATUS command
The TYPEC_STATUS command will be deprecating the informational return
from the USB_PD_CONTROL host command.  It brings over the enabled flags
from that command which are used by the kernel, but leaves behind the
role flags in favor of straightforward enum values.  Cable information
is excluded as it is redundant with the discovery return.  Information
about the mux state is also added for convenience.

Additionally, this moves enums and defines which are a part of our
overall pd_* API to the ec_commands.h file to ensure consumers have the
same field values available for interpretation as the EC.

BRANCH=None
BUG=b:167700356
TEST=on waddledoo, plug in chargers and dongles and ensure outputs from
"ectool typecstatus <port>" match "ectool usbpd <port>" and "ectool
usbpdmuxinfo"

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ic7afc0b282b88fdb34cb9a6feef22ad913bb4aae