TCPMv2: Add TYPEC_STATUS command
The TYPEC_STATUS command will be deprecating the informational return
from the USB_PD_CONTROL host command.  It brings over the enablement,
role, and connection information from the older command.  Cable specifics
are excluded as they are redundant with the discovery return.  Information
about the mux state is also added for convenience.

Additionally, this moves enums and defines which are a part of our
overall pd_* API to the ec_commands.h file to ensure consumers have the
same field values available for interpretation as the EC.

BRANCH=None
BUG=b:167700356
TEST=on waddledoo, plug in chargers and dongles and ensure outputs from
"ectool typecstatus <port>" match "ectool usbpd <port>" and "ectool
usbpdmuxinfo"

Signed-off-by: Diana Z <dzigterman@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2432452
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Change-Id: Id39de2f48e978f9162cbaf59e274f8d6ab85b81f