TCPMv2: Add more specific Vbus presence levels
For boards which use Vbus ADCs, separate the current VBUS_PRESENT check
into VBUS_SAFE5V and VBUS_SINK_DISCONNECT.  The level for vSafe5V should
be used in any locations looking for "Vbus is present" on transition,
but in order to correctly detect disconnection with a load on Vbus,
vSinkDisconnect is required. TODO statemens have been added for places
where work will be needed to support vSinkDisconnectPD.

For boards detecting Vbus through an external chip, the levels will
likely be indistinguishable due to the chips setting a lower threshold
for disconnect than for connection.

BRANCH=None
BUG=b:168831161
TEST=make -j buildall

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I04014ce54ec162dd9c62f545126d921c6d880741