zork: special-case MPU configuration
The NPCX797WC has RAM regions that break the assumptions about alignment
for MPU configuration, and making the MPU configuration code handle this
sort of layout will require complex changes. As a temporary fix for the
zork firmware branch only, hard-code configurations for code and data
RAM (to be replaced with a more general implementation later).

BUG=b:168732993
TEST=verified MPU configuration succeeds on dalboz
BRANCH=zork

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I2aefed6cc381c1244c7788212e07cf62a3ce31a3