zephyr: add compatible hwtimer module
We need the symbols __hw_clock_source_read64 and __hw_clock_event_get
defined for Zephyr to provide times to the CrOS EC.

The event timer does not make sense for Zephyr code, but we need it
defined to prevent link errors (timerinfo uses it).  Perhaps the
solution to this is to add a new config option (e.g.,
CONFIG_EVENT_TIMER) which can be used to selectively enable/disable
the event timer in the CrOS EC.  But punting this work for now and
just adding a dummy definition.

BUG=b:167590251
BRANCH=none
TEST=compile with follow-up CLs

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I58990a6295625f9c34ec080360470431b46155bd