core: rename atomic_clear to atomic_clear_bit
Change the name of atomic_clear to atomic_clear_bit to make to name more
clear - the function clears only selected bits, but the name may suggest
that it clears the whole variable.

It is done as a part of porting to Zephyr, where atomic_clear_bit clears
only selected bits and atomic_clear zeros the variable.

BUG=b:169151160
BRANCH=none
TEST=buildall

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: I7b0b47959c6c54af40f61bca8d9baebaa0375970