SM5803: Leave PROCHOT comparators on for active ports
The PROCHOT comparators provide feedback for OTG Vbus sourcing, so they
should not be disabled while a port is active. Additionally, they must
be enabled before Vbus is sourced and so are not candidates to move into
the PD connect/disconnect hooks.

BRANCH=None
BUG=b:168931726
TEST=on waddledee, reboot EC with a sinking dongle plugged in and
observe the port begins sourcing again in S0 with no errors

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I70f7d8386c650d91257e5d76f95aa5cef145c5ce