SM5803: Leave Vbus PROCHOT comparator on for active ports
The Vbus PROCHOT comparator provides feedback for OTG Vbus sourcing,
so it should not be disabled while a port is active. Additionally, it must
be enabled before Vbus is sourced and so is not a candidate to move into
the PD connect/disconnect hooks.  It can be disabled on PD disconnect in
chipset off or suspended states to save power after unplug.

BRANCH=None
BUG=b:168931726
TEST=on waddledee, reboot EC with a sinking dongle plugged in and
observe the port begins sourcing again in S0 with no errors.  Suspend
DUT and observe sourcing out continues to work, and remove dongle and
observe all PROCHOT comparators are disabled.

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I70f7d8386c650d91257e5d76f95aa5cef145c5ce