usbc_ppc: inform the PPC of the power role of the connected device
On some of the PPC (e.g. syv682) that support manually force discharge
function doesn't automatically turn off discharge FET
when VBUS meets vSafe0V. The original flow is disabling discharge
on pd_set_power_supply_ready and enabling discharge on
pd_power_supply_reset, and since there is no automatic
turning off the discharge circuit, the FET will still be on
when a SNK device connected. We fix this by informing the PPC on
a device is connected or disconnected so that PPC can control the
FET by requests.

BUG=b:160548079, b:148870148, b:163143427
TEST=make buildall
TEST=TCPMv2: on asurada C0/C1(syv682) and volteer C0(sn5s330)/C1(syv682)
     port, and enable force discharge mode (CL:2423665), plug sink
     and source device and ensure Vconn and Vbus are off within
     tVconnOff and tVBusOff respectively. Plug a DRP hub and then
     plug adapter in it, and it meets tVconnOff and tVbusOff.
TEST=TCPMv1: tested the same steps as above on Asurada.
BRANCH=NONE

Change-Id: I8ed0e18fce2d402ff24fce6bab393cc618dfac09
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2434590
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Tested-by: Eric Herrmann <eherrmann@chromium.org>