TCPMv2: Add framework for per-port events
The AP would like to retrieve events on a per-port basis, so add new
tracking for those events, as well as the capability for non-PD tasks to
retrieve and set those events.  For the moment, the code just clears
events on startup and event flags will be added in subsequent CLs.

BRANCH=None
BUG=b:148816435
TEST=make -j buildall

Signed-off-by: Diana Z <dzigterman@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2432454
Change-Id: I189efdea3e3111305b79c431528988343fa43af9