sm5803: Remove initial VSYS setting in init
The OCPC code will ensure that the first VSYS set point is either the
battery voltage, or if not present, the minimum battery voltage as
specified by the batt_info structure.  There's no need to set an
intermediate value here upon init.  The default regulation set point
will be set by the strapping resistor on the charger IC before FW even
gets a chance to run.

BUG=b:168730125
BRANCH=None
TEST=Build and flash drawcia.  Cutoff battery, wait 15s, plug in
charger on C1, verify that VSYS is not initially set with a bad
voltage.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Ibf7c09679fc132c049b79f7ac070d9b3caed794b