ocpc: Misc fixes for coming out of cutoff
This commit fixes a bug in how OCPC determined which portion of the
charge curve that we are in.  Upon startup, the phase is uninitialized
and we get a charge_request of 0,0.  The code assumed that if the
battery voltage was less than the desired voltage, we'd be in the CV
state. But ignoring a zero desired voltage prevented us from assuming
we're in CV.

BUG=b:168135339
BRANCH=None
TEST=Build and flash drawcia, cutoff the battery, wait 15s, plug in a
charger on C1, verify that the DUT comes out of cutoff.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Id5f36d701c2d4e1608954c898c5bc4f0010dec1d