npcx797xc: explicitly set correct code RAM size
The default value for CONFIG_CODE_RAM_SIZE is based on the RO image size,
but the adjusted image size on these chips is incompatible with MPU
configuration because it is not possible to configure a memory region of
that size and alignment (0x3f000 bytes from 0x10070000) in two MPU
regions.

Set CONFIG_CODE_RAM_SIZE to the size of the actual memory region that is
used (rather than the size of the image that gets loaded into it) so the
MPU can be successfully configured.

BUG=b:169276765
BRANCH=zork
TEST=Dalboz MPU configuration succeeds, requesting the expected region
     for code RAM (0x40000 bytes from 0x10070000).

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I73930f73b3ee55a5922abcba57593209ac751e51