TCPMv2: Fix dereferencing a NULL pointer
Some monitors or dongles may not follow the spec that sends a 'DP
Status' VDM to the Chromebook. The 'DP Status' VDM is supposed to
send from a DP source, not a DP sink. When the Chromebook receives
this message, the dereference of a NULL pointer is triggered and
EC crashes. Should ignore this assignment and leave it NULL so the
Chromebook later NACK's the VDM message.

BRANCH=None
BUG=b:169611334
TEST=Build the image.

Signed-off-by: Wai-Hong Tam <waihong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2437189
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Scott Collyer <scollyer@chromium.org>
Change-Id: I3d8730461d9aa08273314ac6778431a10d1e7c66