charge_state_v2: Move the first battery_get_params to HOOK_INIT
Other tasks read the params like state_of_charge at the beginning of
their tasks. In the existing code, the first battery_get_params is
called at the beginning of the charger task. It is not guaranteed the
battery params are ready. This change moves it to the HOOK_INIT.

BRANCH=None
BUG=b:169453974
TEST=Cut off the battery, checked the first boot has PD-enabled, which
uses the battery state_of_charge param to make the decision.

Change-Id: Ie7bd31ee71fb3d1ea47a31910f0dfa7ac93cbcef
Signed-off-by: Wai-Hong Tam <waihong@google.com>