core/nds32: to avoid race condition
If latter task enters the mutex after line 628(original program),
the global variable of mtx->waiters will be written to 1 and go
to sleep waiting, but the local variable of waiter is still 0
because previous task clears waiter to 0 after getting the
mutex lock. This will cause latter task which cannot be woke
up when previous task leaves the mutex unlock.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: tim <tim2.lin@ite.corp-partner.google.com>
Change-Id: I566a14c1120716af8e9ac99e183a989274a4e72e