TCPMv2: clear TC_FLAGS_PR_IN_PROGRESS if ProtocolError
While doing PRS_SNK_SRC on a DRP hub by unpluging the charger,
if the port-partner can't hold tSrcTransistion, and it might brown out
at any time in PR_SWAP AMS (e.g. brown out and can't reply GoodCRC
to Accept), and this will cause a ProtocolError. In that case,
we should clear the TC_FLAGS_PR_IN_PROGRESS so that the TC sm could
start to monitor the VBUS level and restart the PD state machine.

BUG=b:169113820
TEST=unplug the charger from a DRP hub, and DUT could start to
     source the hub instead of staying at SNK_READY.
BRANCH=none

Change-Id: Iccb85a13731004a214446ce4d915fea58170ea00
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>