anx74xx: Improve logging and reliability
This series of CLs is intended as a hotfix to improve stability and
compliance of DisplayPort Alternate Mode on EVE.

The HPD block on ANX3429 is unpowered when in LPM state. By default
this mode is a 44k Input Pulldown. However circuit considerations may
mean this is insufficient to signal HPD disconnect to GPU.

Improve logging for diagnostic purposes, and harden driver code.

BRANCH=firmware-eve-9584.B
BUG=b:167602250
TEST=Flash EVE, initiate Type-C/PD disconnect due to:
        (a) LPM entry
        (b) VBUS threshold loss
        (c) DP ExitMode
     Verify HPD reception by GPU and SSUSB mux state.
Signed-off-by: Nathan K <nkolluru@google.com>

Change-Id: I2fab19d540f8b5c75e69d9ee774f12e5dda1f000