npcx psl: reject hibernate request with timeout
Some Nuvoton NPCX chips support PSL (power state logic) to put the
chip in a very low power state. When this feature is utilized, the
alarm feature is off, thus it is not possible to hibernate for a
specific period of time. Reject CLI requests to do so.

BRANCH=none
BUG=b:144427579
TEST=verified non-zero timeout is reject on volteer

Change-Id: I17537b4724f03bd8f58ee81b80e374629c8b9b88
Signed-off-by: Caveh Jalali <caveh@chromium.org>