core/nds32: add Zephyr compatible atomic functions
Add atomic functions with prototypes equal to the ones in Zephyr.

It is done as a part of porting to Zephyr, the next step is to use in
the code atomic_* instead of deprecated_atomic_*.

Some atomic functions in Zephyr return a value e.g. atomic_add - it
returns the value of the variable before the add operation.
Adjust the atomic functions' code to support that.

GCC builtin functions are not used, because those functions are not
available for nds32.

BUG=b:169151160
BRANCH=none
TEST=buildall

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: I9e181d3c64117b30a08179f84c16162ab70b25fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2444992
Reviewed-by: Jett Rink <jettrink@chromium.org>
Reviewed-by: Dino Li <Dino.Li@ite.com.tw>
Signed-off-by: Keith Short <keithshort@chromium.org>