Shutdown on power button in recovery
WIP & POC

Currently the system shuts down when the power button is pressed
in the recovery mode. EFS2 has to hook this shut down to sysjump
to RW for the next (normal) boot.

This hook can interfer with other warm reboots.

Instead of hooking on shutdown, this patch makes EC reset on a
power button press in the recovery mode.

BUG=b:166787955,b:167996216,chromium:1045209
BRANCH=None
TEST=Puff. Need more.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I51338f94c4f554cd84ed3638b5b3ff21d4685fe6