core/riscv-rv32i: add Zephyr compatible atomic functions
Add atomic functions with prototypes equal to the ones in Zephyr.

It is done as a part of porting to Zephyr, the next step is to use in
the code atomic_* instead of deprecated_atomic_*.

Some atomic functions in Zephyr return a value e.g. atomic_add - it
returns the value of the variable before the add operation.

Use GCC builtin functions instead of the custom asm code in order be
sure there are no bugs and be consistent with other architectures.

atomic_read_add/sub work the same as atomic_add/sub.
atomic_read_add/sub should be replaced and deleted in the next step -
use atomic_* instead of deprecated_atomic_*.

BUG=b:169151160
BRANCH=none
TEST=buildall

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: I1f4f4ef26acb9e388da75125a4dd8e2f0dba9fe8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2444995
Reviewed-by: Jett Rink <jettrink@chromium.org>
Signed-off-by: Keith Short <keithshort@chromium.org>