EFS2: Reset EC when system shuts down to G3
Currently, in recovery mode, EC jumps to RW when the system shuts down
to S5. This allows Chromeboxes to boot in normal mode with PD enabled
in RO, which isn't a desired behavior.

This patch makes EC hard-reset when it's running a RO copy and enters
hard off.

It also makes EC hard-reset when it's running a RO copy and the power
button is pressed before it transitions to G3.

Tried:
1. Boot to recovery screen.
2. Press PB to shutdown.

Then, observed:
1. System resets in 10 seconds.
2. EC jumps to RW by EFS2.
3. System stays off.

Also tried:
1. Boot to recovery screen.
2. Press PB to shutdown.
3. Immediately press PB again.

Then, observed:
1. System resets.
2. EC jumps to RW by EFS2.
3. System boots.

Repeated the above steps with/without AC on Puff and Trembyle.

BUG=b:166787955,b:167996216,chromium:1045209
BRANCH=None
TEST=See the description above.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: Ifdf9438fdf88a0c0b5d28a65a5249502848b054f