TCPMv2: Allow UP_PSwap to be disabled
When we are attached as a source and our partner
has unconstrained power available it is in our
best interest to try to PR_Swap to become the SNK.
In some test scenarios, this defeats what the test
is looking for in a PASS execution. So giving the
option to be able to turn off this by using a new
pd command:

    pd up_pswap [0|1]

Calling this without a parameter will display the
current setting. Providing 0 will DISABLE or turn
the mechanism off. Providing 1, which is the power
on default, will ENABLE or turn the mechanism on.

BUG=b:169802842
BRANCH=zork
TEST=FAFT firmware_PDTrySrc with up_pswap disabled

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: Ic5e4ac1584dd5830d1ed513b664dfbaa1fbf0f70