zephyr: detect the module location in a 2.3/2.4 compatible manner
Most zephyr modules choose to put CMakeLists.txt at the root of the
module, but we instead decided to put ours at zephyr/CMakeLists.txt to
keep all CMake contained within the zephyr/ directory.

So a change in 2.4 came to bite us.  ZEPHYR_CURRENT_MODULE_DIR will
now be set to the base of the module, not the directory where the
CMakeLists.txt is located.

Update the code to probe for the 2.3/2.4 difference.

BUG=b:170268298
BRANCH=none
TEST=compile zephyr for posix-ec on 2.3 and 2.4 (using zmake; build
     instructions and WIP CL can be found at go/zmake)

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: Ia6da8451bd7d25aafe7cf42066d202ead208fa7c