TCPMv2: Enable PD comm when the battery level is enough
This is for CONFIG_SYSTEM_UNLOCKED and not CONFIG_VBOOT_EFS2.
The PD comm is disabled on startup if the battery level is low.

Some batteries take much longer time to report its SOC. As a
result, the PD comm has no way back to enable. This CL adds
a HOOK when the battery SOC change and enables PD comm when
the battery level is enough.

BRANCH=None
BUG=b:169453974
TEST=Hacked to fake the battery low on startup and restored
it afterward. Checked the PD comm was enabled later.

Change-Id: I332f76b4fc635c15d29b8e98aec9cae4eb605246
Signed-off-by: Wai-Hong Tam <waihong@google.com>